test_app/requirements.txt
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
asgiref==3.3.1
djangorestframework==3.12.2
psycopg2==2.8.6
PyJWT==2.0.1
